Implement ConsoleMedia
ConsoleMedia is specifically trimmed to support console printing.
But it does not print directly to a console.
Instead it collects lines for each value printed and returns those
as a single String when asked.

This fixes #21
