+++ pom.xml
@@ -56,6 +56,7 @@
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <versions.build.findbugs-maven-plugin>3.0.4</versions.build.findbugs-maven-plugin>
         <versions.build.jacoco-maven-plugin>0.7.9</versions.build.jacoco-maven-plugin>
+        <versions.build.junit-surefire-provider>1.0.0-M4</versions.build.junit-surefire-provider>
         <versions.build.maven-checkstyle-plugin>2.17</versions.build.maven-checkstyle-plugin>
         <versions.build.maven-compiler-plugin>3.6.0</versions.build.maven-compiler-plugin>
         <versions.build.maven-deploy-plugin>2.8.2</versions.build.maven-deploy-plugin>
@@ -66,7 +67,7 @@
         <versions.build.nexus-staging-maven-plugin>1.6.8</versions.build.nexus-staging-maven-plugin>
         <versions.build.versions-maven-plugin>2.3</versions.build.versions-maven-plugin>
         <versions.dependencies.json>[1.1.0-M1,1.2.0)</versions.dependencies.json>
-        <versions.dependencies.junit>5.0.0-M3</versions.dependencies.junit>
+        <versions.dependencies.junit>5.0.0-M4</versions.dependencies.junit>
         <versions.dependencies.paper>[2.3.1,3.0.0)</versions.dependencies.paper>
         <versions.java>1.8</versions.java>
     </properties>
@@ -186,7 +187,7 @@
                     <dependency>
                         <groupId>org.junit.platform</groupId>
                         <artifactId>junit-platform-surefire-provider</artifactId>
-                        <version>1.0.0-M3</version>
+                        <version>${versions.build.junit-surefire-provider}</version>
                     </dependency>
                     <dependency>
                         <groupId>org.junit.jupiter</groupId>
